AugmentorAddons

Documentation for AugmentorAddons.

AugmentorAddons.PadBorders

AugmentorAddons.RCutOut

AugmentorAddons.PadBorders — Type

PadBorders(border, lo, up)
PadBorders(border, bo)
PadBorders(border, bo...)

The argument border is a symbol or value:

:replicate repeat edge values to infinity

:circular image edges "wrap around"

:symmetric the image reflects relative to a position between pixels

:reflect the image reflects relative to the edge itself

a numeric or color type value: for example RGB(0, 0, 0) will pad a RGB image with black

lo and up are tuples:

lo number to extend by on the lower edge for each dimension

up number to extend by on the upper edge for each dimension

If a single tuple (or an integer serie) is supplied, the padding is symmetric.

If PadBorders is used to process an image batch, you must define lo and hi as tuples with the batch dimension set to 0.

Example

# Pad the testpattern from Augmentor with black
julia> augment(PadBorders(RGB(0, 0, 0), (50, 100))) |> plot

# Pad the testpattern from Augmentor with circular borders
julia> augment(PadBorders(:circular, 50, 50)) |> plot

# Pad a fake image batch with 0
julia> augment(rand(32, 32, 3, 16), PadBorders(0, 8, 8, 0, 0))[:,:,1,1] |> heatmap

See also Pad, and Fill.

AugmentorAddons.RCutOut — Type

RCutOut(m; rng = GLOBAL_RNG)
RCutOut(m...; rng = GLOBAL_RNG)

This operation applies a mask of zeros to a random location within images. m is a tuple or integers defining the mask size.

Example

julia> augment(RCutOut((100, 100))) |> plot

julia> augment(RCutOut(100, 100)) |> plot